Replace lastRequestedFrame variable with grabber's lastFrame()
Replace the lastRequestedFrame member variable in MainWindow
that tracked the frame where we can jump if the script is
updated with video frame grabber's lastFrame() method.

Also fix a possibly misbehaving setVideoSource in the grabber.

We also had to change a couple of the queued connection requests
to the frame grabber with blocking calls so that the lastFrame()
functions return the correct frame number without explicit increments.

The queued connection in seekSlider_valueChanged was changed to
a blocking call so that the UI doesn't change out of sync (too early).